fix: removed access cookie
