+++ apps/artemis-web/middleware/auth.ts
@@ -1,10 +1,11 @@
 import nextConnect from 'next-connect';
 import passport from '../lib/passport';
 import database from './database';
+import session from '../middleware/session';
 
 const auth = nextConnect()
   .use(database)
-  // .use(session)
+  .use(session)
   .use(passport.initialize())
   .use(
     passport.session({
